Pulling in all changes from master to android-api branch
The Android-API branch was missing a large amount of changes
from the master branch, so this commit copies/pastes in every
file from master into this branch.

Reviewed-on: https://gerrit.iotivity.org/gerrit/834
Tested-by: Erich Keane <erich.keane@intel.com>

+++ service/things-manager/sdk/src/GroupManager.h
@@ -111,7 +111,7 @@ public:
     OCStackResult executeActionSet(std::shared_ptr< OCResource > resource,
             std::string actionsetName, PostCallback cb);
     OCStackResult executeActionSet(std::shared_ptr< OCResource > resource,
-            std::string actionsetName, unsigned int delay, PostCallback cb);
+            std::string actionsetName, long int delay, PostCallback cb);
     OCStackResult cancelActionSet(std::shared_ptr< OCResource > resource,
             std::string actionsetName, PostCallback cb);
     OCStackResult getActionSet(std::shared_ptr< OCResource > resource, std::string actionsetName,
